- Renamed my package. - Bug fix: quiz can now handle different cases properly.
