+++ src/data/Users.csv
@@ -1,4 +1,5 @@
 username,password,isAdmin,flashcardSetIds
 jempio,ilovejempio,false
 ducas,badpassword,false
+Anthony,123456,false,2
 testUser,password,false,0,1
