+++ src/main/java/MainPage/FlashcardSetDataPanel.java
@@ -4,8 +4,8 @@ import EditorMainPage.EditorMainPage;
 import dataAccess.DBGateway;
 import delete_flashcardset_use_case.*;
 import login_and_signup_use_case.UserLoginResponseModel;
-import quizUseCase.*;
-import quizUseCase.screens.QuizSettingsScreen;
+import quiz_use_case.*;
+import quiz_use_case.screens.QuizSettingsScreen;
 import studyMode.*;
 import studyMode.screens.StudySettingsScreen;
