+++ src/main/java/quiz_use_case/GUI/Slider.java
@@ -1,4 +1,4 @@
-package quizUseCase.GUI;
+package quiz_use_case.GUI;
 
 import javax.swing.*;
 import javax.swing.event.ChangeListener;
