+++ src/main/java/quiz_use_case/QuizInteractor.java
@@ -1,4 +1,4 @@
-package quizUseCase;
+package quiz_use_case;
 
 import dataAccess.DBGateway;
 import entities.*;
