+++ src/main/java/quiz_use_case/QuizOutputBoundary.java
@@ -1,4 +1,4 @@
-package quizUseCase;
+package quiz_use_case;
 
 import java.util.ArrayList;
