+++ src/main/java/quiz_use_case/QuizSettingsRequestModel.java
@@ -1,4 +1,4 @@
-package quizUseCase;
+package quiz_use_case;
 
 /**
  * Quiz Settings Request Model.
